move modern BottomModal
