+++ src/screens/WalletConnect/DisconnectModal/WalletConnectDisconnectModal.js
@@ -22,7 +22,7 @@ import * as React from 'react';
 import { useTranslationWithPrefix } from 'translations/translate';
 
 // Components
-import BottomModal from 'components/modern/BottomModal';
+import BottomModal from 'components/layout/BottomModal';
 import Button from 'components/core/Button';
 import Image from 'components/Image';
 import Toast from 'components/Toast';
